requirements.txt
Misskey.py
gspread